requirements.txt
asn1crypto==0.24.0
backcall==0.1.0
beautifulsoup4==4.8.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
filelock==3.0.12
glob2==0.7
h5py==2.10.0
idna==2.8
imageio==2.8.0
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
Jinja2==2.10.1
joblib==0.14.1
kiwisolver==1.2.0
libarchive-c==2.8
lief==0.12.3
MarkupSafe==1.1.1
matplotlib==3.2.1
networkx==2.4
numpy==1.21.0
olefile==0.46
pandas==1.0.3
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.2
pkginfo==1.5.0.1
prompt-toolkit==2.0.9
protobuf==3.11.3
psutil==5.6.6
ptyprocess==0.6.0
pycosat==0.6.3
pycparser==2.19
Pygments==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.7
PySocks==1.7.0
python-dateutil==2.8.1
pytz==2019.2
PyWavelets==1.1.1
PyYAML==5.1.2
requests==2.22.0
ruamel-yaml==0.17.21
scikit-image==0.16.2             
scikit-learn==0.22.2.post1       
scipy==1.10.1 
six==1.12.0
sklearn==0.0
soupsieve==1.9.2
tensorboardX==2.0
tqdm==4.32.1
traitlets==4.3.2
urllib3==1.24.2
wcwidth==0.1.7
opencv-python==4.7.0.72